pyproject.toml
[project]
name = "jcatch-plugin"
version = "0.1.0"
description = "JAV video metadata scraper plugin"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "selenium>=4.0.0",
    "webdriver-manager>=4.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
jcatch-plugin = "jcatch_plugin.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
